Fixed linting issues and refactored some of our code to settingsEditor.ts Passing UI smoke tests Passing manual QA tests Ready to merge into master @TejDNagaraja Issue #71836
